@@ -14,79 +15,80 @@
     },
     "devDependencies": {
         "@pixelandtonic/craftui": "~0.3.8",
-        "@rollup/plugin-node-resolve": "^13.3.0",
-        "@vitejs/plugin-legacy": "^1.8.2",
-        "@vitejs/plugin-vue": "^2.3.3",
-        "autoprefixer": "^10.4.7",
+        "@vitejs/plugin-legacy": "^5.2.0",
+        "@vitejs/plugin-vue": "^4.5.0",
+        "autoprefixer": "^10.4.16",
         "clean-webpack-plugin": "^4.0.0",
         "craftcms-sass": "^3.5.7",
-        "eslint": "^8.19.0",
+        "eslint": "^8.54.0",
         "eslint-import-resolver-alias": "^1.1.2",
-        "eslint-plugin-import": "^2.26.0",
-        "eslint-plugin-vue": "^9.2.0",
-        "eslint-webpack-plugin": "^3.2.0",
+        "eslint-plugin-import": "^2.29.0",
+        "eslint-plugin-vue": "^9.19.1",
+        "eslint-webpack-plugin": "^4.0.1",
         "laravel-mix": "^6.0.49",
         "laravel-mix-clean": "^0.1.0",
         "laravel-mix-polyfill": "^3.0.1",
         "lodash-es": "^4.17.21",
-        "mini-css-extract-plugin": "^2.6.1",
-        "postcss": "^8.4.14",
+        "mini-css-extract-plugin": "^2.7.6",
+        "postcss": "^8.4.31",
         "postcss-color-function": "^4.1.0",
-        "postcss-custom-properties": "^12.1.8",
+        "postcss-custom-properties": "^13.3.2",
         "rollup-plugin-analyzer": "^4.0.0",
-        "sass": "^1.53.0",
-        "sass-loader": "^13.0.2",
-        "vite": "^2.9.13",
-        "vite-plugin-eslint": "^1.6.1",
+        "sass": "^1.69.5",
+        "sass-loader": "^13.3.2",
+        "vite": "^5.0.4",
+        "vite-plugin-eslint": "^1.8.1",
         "vite-plugin-imagemin": "^0.6.1",
-        "vuex": "^4.0.2"
+        "vite-plugin-node-polyfills": "^0.17.0",
+        "vuex": "^4.1.0"
     },
     "dependencies": {
-        "@formkit/themes": "^1.0.0-beta.9-4ee1e0a",
-        "@formkit/vue": "^1.0.0-beta.9-4ee1e0a",
-        "@fortawesome/fontawesome-svg-core": "^6.1.1",
-        "@fortawesome/free-regular-svg-icons": "^6.1.1",
-        "@fortawesome/free-solid-svg-icons": "^6.1.1",
-        "@fortawesome/vue-fontawesome": "^3.0.0-5",
+        "@formkit/pro": "^0.118.1",
+        "@formkit/themes": "^1.3.2",
+        "@formkit/vue": "^1.3.2",
+        "@fortawesome/fontawesome-svg-core": "^6.5.0",
+        "@fortawesome/free-regular-svg-icons": "^6.5.0",
+        "@fortawesome/free-solid-svg-icons": "^6.5.0",
+        "@fortawesome/vue-fontawesome": "^3.0.5",
         "@rynpsc/dialog": "^5.0.2",
-        "@tiptap/core": "^2.0.0-beta.182",
-        "@tiptap/extension-floating-menu": "^2.0.0-beta.56",
-        "@tiptap/extension-focus": "^2.0.0-beta.45",
-        "@tiptap/extension-highlight": "^2.0.0-beta.35",
-        "@tiptap/extension-image": "^2.0.0-beta.30",
-        "@tiptap/extension-link": "^2.0.0-beta.43",
-        "@tiptap/extension-subscript": "^2.0.0-beta.13",
-        "@tiptap/extension-superscript": "^2.0.0-beta.13",
-        "@tiptap/extension-text-align": "^2.0.0-beta.31",
-        "@tiptap/extension-underline": "^2.0.0-beta.25",
-        "@tiptap/starter-kit": "^2.0.0-beta.191",
-        "@tiptap/vue-3": "^2.0.0-beta.96",
-        "@yaireo/tagify": "^4.12.0",
+        "@tiptap/core": "^2.1.12",
+        "@tiptap/extension-floating-menu": "^2.1.12",
+        "@tiptap/extension-focus": "^2.1.12",
+        "@tiptap/extension-highlight": "^2.1.12",
+        "@tiptap/extension-image": "^2.1.12",
+        "@tiptap/extension-link": "^2.1.12",
+        "@tiptap/extension-subscript": "^2.1.12",
+        "@tiptap/extension-superscript": "^2.1.12",
+        "@tiptap/extension-text-align": "^2.1.12",
+        "@tiptap/extension-underline": "^2.1.12",
+        "@tiptap/starter-kit": "^2.1.12",
+        "@tiptap/vue-3": "^2.1.12",
+        "@yaireo/tagify": "^4.17.9",
         "click-outside-vue3": "^4.0.1",
         "expression-language": "^1.1.4",
-        "flat": "^5.0.2",
+        "flat": "^6.0.1",
         "flatpickr": "^4.6.13",
-        "floating-vue": "^2.0.0-beta.17",
+        "floating-vue": "^2.0.0-beta.24",
         "formdata-polyfill": "^4.0.10",
-        "friendly-challenge": "^0.9.11",
+        "friendly-challenge": "^0.9.14",
         "hostic-dom": "^0.8.7",
         "install": "^0.13.0",
-        "intl-tel-input": "^17.0.16",
+        "intl-tel-input": "^18.2.1",
         "js-cookie": "^3.0.5",
         "locutus": "^2.0.16",
-        "markdown-it": "^13.0.1",
-        "md5-hex": "^4.0.0",
-        "mobile-drag-drop": "^3.0.0-beta.0",
-        "npm": "^8.13.2",
+        "markdown-it": "^13.0.2",
+        "md5-hex": "^5.0.0",
+        "mobile-drag-drop": "^3.0.0-rc.0",
+        "npm": "^10.2.4",
         "payment": "^2.4.6",
         "pell": "^1.0.6",
         "places.js": "^1.19.0",
-        "qs": "^6.11.0",
-        "sanitize-html": "^2.7.0",
-        "signature_pad": "^4.0.5",
+        "qs": "^6.11.2",
+        "sanitize-html": "^2.11.0",
+        "signature_pad": "^4.1.7",
         "sortablejs": "^1.15.0",
         "tippy.js": "^6.3.7",
-        "vue": "^3.2.37",
+        "vue": "^3.3.9",
         "vue-final-modal": "^3.4.3",
         "vue-mitt": "^0.0.1",
         "vue-multiselect": "^3.0.0-alpha.2",
